# backend/docs/swagger.yaml
definitions:
  models.AuthRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  models.ProjectUpsertDTO:
    properties:
      description:
        type: string
      name:
        type: string
      teamId:
        type: string
    required:
    - name
    - teamId
    type: object
  models.TeamUpsertDTO:
    properties:
      description:
        type: string
      name:
        type: string
    required:
    - name
    type: object
  models.UserCreateDTO:
    properties:
      email:
        type: string
      firstName:
        type: string
      lastName:
        type: string
      password:
        type: string
      role:
        type: string
      teamId:
        type: string
      username:
        type: string
    required:
    - email
    - firstName
    - lastName
    - password
    - role
    - username
    type: object
  models.UserUpdateDTO:
    properties:
      email:
        type: string
      firstName:
        type: string
      lastName:
        type: string
      role:
        type: string
      teamId:
        type: string
      username:
        type: string
    required:
    - email
    - firstName
    - lastName
    - role
    - username
    type: object
info:
  contact: {}
paths:
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticates a user and returns a JWT token
      parameters:
      - description: Login credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.AuthRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Authentication failed
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: User login
      tags:
      - auth
  /api/auth/register:
    post:
      consumes:
      - application/json
      description: Registers a new user
      parameters:
      - description: User registration info
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.UserCreateDTO'
      produces:
      - application/json
      responses:
        "201":
          description: User registered successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request payload
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to register user
          schema:
            additionalProperties: true
            type: object
      summary: Register user
      tags:
      - auth
  /api/projects:
    get:
      description: Get a list of all projects
      produces:
      - application/json
      responses:
        "200":
          description: List of projects
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to list projects
          schema:
            additionalProperties: true
            type: object
      summary: List projects
      tags:
      - projects
    post:
      consumes:
      - application/json
      description: Create a new project
      parameters:
      - description: Project info
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.ProjectUpsertDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Project created successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request payload
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to create project
          schema:
            additionalProperties: true
            type: object
      summary: Create project
      tags:
      - projects
  /api/projects/{id}:
    delete:
      description: Delete a project by its ID
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Project deleted successfully
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to delete project
          schema:
            additionalProperties: true
            type: object
      summary: Delete project
      tags:
      - projects
    get:
      description: Get details of a project by its ID
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Project details
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Get project by ID
      tags:
      - projects
    put:
      consumes:
      - application/json
      description: Update an existing project
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Project info
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.ProjectUpsertDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Project updated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request payload
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to update project
          schema:
            additionalProperties: true
            type: object
      summary: Update project
      tags:
      - projects
  /api/teams:
    get:
      description: Get a list of all teams
      produces:
      - application/json
      responses:
        "200":
          description: List of teams
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to list teams
          schema:
            additionalProperties: true
            type: object
      summary: List teams
      tags:
      - teams
    post:
      consumes:
      - application/json
      description: Create a new team
      parameters:
      - description: Team info
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.TeamUpsertDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Team created successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request payload
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to create team
          schema:
            additionalProperties: true
            type: object
      summary: Create team
      tags:
      - teams
  /api/teams/{id}:
    delete:
      description: Delete a team by its ID
      parameters:
      - description: Team ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Team deleted successfully
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to delete team
          schema:
            additionalProperties: true
            type: object
      summary: Delete team
      tags:
      - teams
    get:
      description: Get details of a team by its ID
      parameters:
      - description: Team ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Team details
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Get team by ID
      tags:
      - teams
    put:
      consumes:
      - application/json
      description: Update an existing team
      parameters:
      - description: Team ID
        in: path
        name: id
        required: true
        type: string
      - description: Team info
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.TeamUpsertDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Team updated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request payload
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to update team
          schema:
            additionalProperties: true
            type: object
      summary: Update team
      tags:
      - teams
  /api/teams/{teamID}/users:
    get:
      description: Get a list of users by team ID
      parameters:
      - description: Team ID
        in: path
        name: teamID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of users by team
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to list users by team ID
          schema:
            additionalProperties: true
            type: object
      summary: List users by team ID
      tags:
      - users
  /api/users:
    get:
      description: Get a list of all users
      produces:
      - application/json
      responses:
        "200":
          description: List of users
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to list users
          schema:
            additionalProperties: true
            type: object
      summary: List users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Create a new user
      parameters:
      - description: User info
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.UserCreateDTO'
      produces:
      - application/json
      responses:
        "201":
          description: User created successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request payload
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to create user
          schema:
            additionalProperties: true
            type: object
      summary: Create user
      tags:
      - users
  /api/users/{id}:
    delete:
      description: Delete a user by their ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User deleted successfully
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to delete user
          schema:
            additionalProperties: true
            type: object
      summary: Delete user
      tags:
      - users
    get:
      description: Get details of a user by their ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User details
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Get user by ID
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update an existing user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: User info
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.UserUpdateDTO'
      produces:
      - application/json
      responses:
        "200":
          description: User updated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request payload
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to update user
          schema:
            additionalProperties: true
            type: object
      summary: Update user
      tags:
      - users
swagger: "2.0"